refactor: move eve runtime under leaf internals
