+++ apps/leaf/src/internal/agentRuntime/eve/classifyParkedInput.ts
@@ -1,5 +1,5 @@
-import { normalizeToolName } from "../../agent/tools/toolPolicy.js";
-import { WAITING_FOR_INPUT_MESSAGE } from "../../ui/messages.js";
+import { normalizeToolName } from "../../../agent/tools/toolPolicy.js";
+import { WAITING_FOR_INPUT_MESSAGE } from "../../../ui/messages.js";
 import { type EveInputRequest, textForInputRequests } from "./events.js";
 
 /** A gated write the resumed turn parked on after the answered one. */
