+++ apps/leaf/src/internal/agentRuntime/eve/embeddedServer.ts
